add: dictionaries and lang pages
